refactor(agent): split semantic send models
